Disable profiling by default
Profiling is a very CPU hungry task. Information leaked
by pprof alone is limited, an attacker may be able to
learn about the target binary internals such as its
functions or stack traces which may be helpful when
dealing with an unfamiliar binary. Other than that,
it may be possible to result in DoS of the target
through pprof.

https://www.twistlock.com/labs-blog/recent-kubernetes-vulnerability-exposed-pprof-interface/